layan

+++ newAppointment.php
@@ -47,31 +47,44 @@
        <p><a href="bookAppointment.php"><button type="submit">Book</button></a>
     
     </form>
-    <?php
-             if ($_SERVER["REQUEST_METHOD"] == "POST") {
-                 if ( !( $database = mysqli_connect( "localhost", "root", "" ) ) )
-                    die( "<p>Could not connect to database</p>" );
 
-                 if ( !mysqli_select_db( $database, "Pet_care") )
-                    die( "<p>Could not open URL database</p>" );
 
-                    $name =  $_POST["PetName"];
-                    $note =   $_POST["note"];
-                    $owner_email = $_POST['owner_email'];
-                    $service= $_GET['service'];
-                    $date = $_GET['date'];
-                    $time = $_GET['time'];
-                    $id =  $_GET['id'];
-                    mysqli_query($database ,"DELETE FROM available_appointments WHERE appointment_id='".$id."'");
-                 $query="INSERT INTO appointments_requests (pet_name, service, date , time ,owner_email , note ) VALUES ('".$name."','".$service."','".$date."','".$time."','".$owner_email."','".$note."');";
-                 $result=mysqli_query($database, $query);
-                 if($result)
-                 header("location:AppointmentRequest.php");
-                
 
-               
-             }
+    <?php 
+       if ($_SERVER["REQUEST_METHOD"] == "POST") {
+        if ( !( $database = mysqli_connect( "localhost", "root", "" ) ) )
+           die( "<p>Could not connect to database</p>" );
+
+        if ( !mysqli_select_db( $database, "Pet_care") )
+           die( "<p>Could not open URL database</p>" );
+
+           $owner_email = $_POST['owner_email']; 
+           if(filter_var(  $owner_email , FILTER_VALIDATE_EMAIL))
+           {
+           
+            $name =  $_POST["PetName"];
+            $note =   $_POST["note"];
+            $owner_email = $_POST['owner_email'];
+            $service= $_GET['service'];
+            $date = $_GET['date'];
+            $time = $_GET['time'];
+            $id =  $_GET['id'];
+            
+            mysqli_query($database ,"DELETE FROM available_appointments WHERE appointment_id='".$id."'");
+         $query="INSERT INTO appointments_requests (pet_name, service, date , time ,owner_email , note ) VALUES ('".$name."','".$service."','".$date."','".$time."','".$owner_email."','".$note."');";
+         $result=mysqli_query($database, $query);
+         if($result)
+         header("location:AppointmentRequest.php");
+        
+          } 
+          else {
+            echo(" invalid email");
+
+          }
+        }
         ?>
+        
+        
     </div>
